reorganise test lib, allow debug of exec operation, remove sourcemap related asserts
