+++ test/cases/common/test/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const sequence = require('promise-compose');
+
+const all = (...tests) => (...elements) =>
+  sequence(...tests.map((test) => test(...elements)));
+
+module.exports = Object.assign(
+  {all},
+  require('./valid'),
+  require('./invalid')
+);
