+++ test/index.js
@@ -11,7 +11,7 @@ const {init, layer, cwd, fs, env, meta, exec} = require('test-my-cli');
 const {assign} = Object;
 
 const {assertExitCodeZero} = require('./lib/assert');
-const {testBase} = require('./cases/common/tests');
+const {testBase} = require('./cases/common/test');
 
 // tests are located in resolve-url-loader package which might differ from package under test
 const PLATFORMS_DIR = compose(normalize, join)(__dirname, '..', 'packages', 'resolve-url-loader', 'test');
@@ -116,6 +116,9 @@ filterTests()
           },
           layer: {
             keep: (process.env.KEEP === 'true')
+          },
+          exec: {
+            debug: (process.env.DEBUG === 'exec')
           }
         }),
         ...filterTests(platform).map(engine =>
