+++ test/lib/assert/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = Object.assign(
+  {},
+  require('./selective'),
+  require('./exec'),
+  require('./file'),
+  require('./stream')
+);
