+++ test/lib/assert/util.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const {join, relative} = require('path');
+
+exports.subdir = ({root, cwd, env: {OUTPUT}}) =>
+  relative(root, join(cwd, OUTPUT));
